new layout setup so login has its own layout
