+++ apps/web/app/(system)/dashboard/inner/page.tsx
@@ -1,0 +1,15 @@
+'use client'
+
+// app/dashboard/page.tsx
+
+import React from 'react';
+
+export default function innerPage() {
+    return (
+      <>
+        <div style={{ background: 'red', height: '700px' }}>
+          Inner Dashboard Page
+        </div>
+      </>
+    );
+  }
