+++ apps/web/app/(system)/dashboard/page.tsx
@@ -1,0 +1,17 @@
+'use client'
+
+// app/dashboard/page.tsx
+
+import React from 'react';
+
+export default function DashboardPage() {
+    return (
+      <>
+        <div style={{ background: 'red', height: '700px' }}>
+          Dashboard Content
+        </div>
+      </>
+    );
+  }
+
+
