+++ apps/web/app/(system)/dokumen-lelang/page.tsx
@@ -1,0 +1,15 @@
+'use client'
+
+// app/dashboard/page.tsx
+
+import React from 'react';
+
+export default function DokumenLelangForm() {
+    return (
+      <>
+        <div style={{ background: 'red', height: '700px' }}>
+            Lelang Content
+        </div>
+      </>
+    );
+  }
